+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## CHANGELOG (5/02/2021 3:39 pm)
+## CHANGELOG (5/04/2021 5:31 pm)
 
-1. changed post time display method from toString() to toLocateDateString().
-2. fixed post in pathname not being shown (there is a still a better way).
+1. fixed post in pathname not being shown.
+2. moved all components to their own files.
+3. fixed the findComment function to avoid fetching comments every time.
+4. added react-router-dom and made it work.
+5. previous post's comments dont get rendered until new comments get loaded when the post is changed.